allow passing in TLS connector
We've been through a number of different possible rustls connectors,
and we're currently using one (rustls-platform-verifier) that
doesn't have feature flags in livekit or tungstenite yet.

Instead of requiring that every transitive dependency supports the
rustls feature flags we need, it's much easier to pass only the
__rustls-tls feature flag and pass the Connector that we create
with our configuration down the stack.