move starting/stopping GLib event loop into libwebrtc crate (#798)
instead of the screensharing example. This is an implementation
detail that users of the library should not need to be concerned
about. It is likely this library will be used in cross platform
where the developers might not be very familiar with Linux and
could overlook the need to run a GLib event loop if they have to
handle it on their own.